+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [5.5.0] - 2026-01-02
+
+### Added
+
+- Added support of some variants of `ImageBuffer` type
+  form the `image` crate ([#57](https://github.com/Cykooz/fast_image_resize/issues/57)).
+
 ## [5.4.0] - 2025-11-28
 
 ### Added